postgresql: transition to metapackage
We are now shipping PostgreSQL with multiple versions. This involves a
packageing strategy change, where the `postgresql' package itself no
longer holds the server it self, and instead is a metapackage that
points to the current latest major release, e.g. `postgresql-17'.

Also we are splitting PostgreSQL into runtime and server packages, so
there's no unnecessary things installed to the system (Qt depends on
PostgreSQL runtime libraries, but most of the time users don't need to
run the PostgreSQL server itself.)

This metapackage also contains symlinks of executables, libraries and
manual pages to the current major release.

This metapackage also will notify the user about the transition. If the
user is running a PostgreSQL server right now, they will be able to shut
down the services that are using their PostgreSQL server before the
transition happens. A shutdown is required to complete this transition,
if the previous PostgreSQL is running, to gracefully write out the data,
to aid the possible migration. It will also tell the user to install the
old release of PostgreSQL if they want to either continue runnig or
migrate to the current major release.